全局配置，不再自动删除无效的配置参数

+++ apps/app_global/configs_manager.py
@@ -21,12 +21,7 @@ class Configs(metaclass=Singleton):
     CACHE_KEY = 'app_global_configs_cache'
 
     @staticmethod
-    def get_configs():
-        cache_key = Configs.CACHE_KEY
-        configs = dj_cache.get(cache_key)
-        if configs:
-            return configs
-
+    def get_configs_no_cache(remove_invalid: bool = False):
         qs = GlobalConfig.objects.all().values('name', 'value')
         configs = {}
         invalid_names = []
@@ -37,7 +32,7 @@ class Configs(metaclass=Singleton):
             else:
                 invalid_names.append(name)
 
-        if invalid_names:
+        if remove_invalid and invalid_names:
             GlobalConfig.objects.filter(name__in=invalid_names).delete()
 
         # 缺少配置
@@ -48,6 +43,16 @@ class Configs(metaclass=Singleton):
                         name=name, defaults={'value': GlobalConfig.value_defaults[name]})
                     configs[name] = obj.value
 
+        return configs
+
+    @staticmethod
+    def get_configs():
+        cache_key = Configs.CACHE_KEY
+        configs = dj_cache.get(cache_key)
+        if configs:
+            return configs
+
+        configs = Configs.get_configs_no_cache(remove_invalid=False)
         dj_cache.set(cache_key, configs, timeout=120)
         return configs
